Deleted the original addItemInfo storyboard & VC. Added new blank AddItemInfo storyboard & VC.
