--- iOS/fleaMarket/TabStoryReferenceSample/SelectFromListTableViewCell.swift
@@ -1,24 +1,0 @@
-//
-//  SelectFromListTableViewCell.swift
-//  fleaMarket
-//
-//  Created by Kero on 2016/6/11.
-//  Copyright © 2016年 ColorKit. All rights reserved.
-//
-
-import UIKit
-
-class SelectFromListTableViewCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
