--- iOS/fleaMarket/TabStoryReferenceSample/ShowDataTableViewCell.swift
@@ -1,28 +1,0 @@
-//
-//  ShowDataTableViewCell.swift
-//  fleaMarket
-//
-//  Created by Kero on 2016/6/11.
-//  Copyright © 2016年 ColorKit. All rights reserved.
-//
-
-import UIKit
-
-class ShowDataTableViewCell: UITableViewCell {
-
-    @IBOutlet weak var leftLabel: UILabel!
-    
-    @IBOutlet weak var rightLabel: UILabel!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
